Rename everything from module to plugin
