+++ src/Plugins/PluginInterface.php
@@ -2,16 +2,16 @@
 
 declare(strict_types=1);
 
-namespace OpenEMR\Modules;
+namespace OpenEMR\Plugins;
 
-interface ModuleInterface
+interface PluginInterface
 {
     /**
      * Returns paths to configuration files that define services for dependency
      * injection. These are merged into the application's container
-     * configuration when the module is active.
+     * configuration when the plugin is active.
      *
-     * @return string[] Module-root relative file paths
+     * @return string[] Plugin-root relative file paths
      */
     public static function getConfigFiles(): array;
 }
