+++ src/Plugins/PluginManager.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace OpenEMR\Modules;
+namespace OpenEMR\Plugins;
 
 use Psr\Container\ContainerInterface;
 use Symfony\Component\Console\Application;
 
-class ModuleManager
+class PluginManager
 {
     public function __construct(
         private Activation\StateProvider $activationState,
@@ -15,15 +15,15 @@ class ModuleManager
     }
 
     /**
-     * @template T of ModuleInterface
+     * @template T of PluginInterface
      * @param class-string<T> $filter
      * @return iterable<class-string<T>>
      */
-    private function getActiveModules(string $filter = ModuleInterface::class): iterable
+    private function getActivePlugins(string $filter = PluginInterface::class): iterable
     {
         // TODO: cache/memoize this data?
-        $moduleData = \OpenEMR\PluginInstaller\GeneratedInstalledPlugins::PLUGIN_DATA;
-        foreach ($moduleData as $package => $data) {
+        $pluginData = \OpenEMR\PluginInstaller\GeneratedInstalledPlugins::PLUGIN_DATA;
+        foreach ($pluginData as $package => $data) {
             if (!$this->activationState->isActive($package)) {
                 continue;
             }
@@ -38,7 +38,7 @@ class ModuleManager
     public function addCommands(Application $cli, ContainerInterface $container): void
     {
         // TODO: Figure out how to make these lazy-loaded
-        foreach ($this->getActiveModules(CliModuleInterface::class) as $bootstrap) {
+        foreach ($this->getActivePlugins(CliPluginInterface::class) as $bootstrap) {
             $commands = $bootstrap::getCommandClasses();
             foreach ($commands as $command) {
                 $cli->addCommand($container->get($command));
